add tests

+++ src/cluster_mgr/src/cli/task/group/remove_group.rs
@@ -103,21 +103,24 @@ impl TaskGroup for RemoveTaskGroup {
         match store.provider().unwrap() {
             StorageProvider::Cassandra => {
                 let cass = store.cassandra.as_ref().unwrap();
-                let host = cass.host.first().unwrap().clone();
-                let task_id = TaskId {
-                    cmd: "remove".to_string(),
-                    task: "drop-keyspace".to_string(),
-                    host: "_local".to_string(),
-                };
-                let cql = format!("DROP KEYSPACE {}", config.deployment.get_keyspace()?);
-                let task = CassandraOpTask::new(task_id.clone(), host, cass.client_port()?, cql);
-                let inst = TaskInstance {
-                    task_input: HashMap::default(),
-                    task: Box::new(task),
-                    task_host: TaskHost::Local,
-                };
-                barrier.push(1);
-                executable.insert(task_id, inst);
+                if cass.external().is_some() {
+                    let host = cass.host.first().unwrap().clone();
+                    let task_id = TaskId {
+                        cmd: "remove".to_string(),
+                        task: "drop-keyspace".to_string(),
+                        host: "_local".to_string(),
+                    };
+                    let cql = format!("DROP KEYSPACE {}", config.deployment.get_keyspace()?);
+                    let task =
+                        CassandraOpTask::new(task_id.clone(), host, cass.client_port()?, cql);
+                    let inst = TaskInstance {
+                        task_input: HashMap::default(),
+                        task: Box::new(task),
+                        task_host: TaskHost::Local,
+                    };
+                    barrier.push(1);
+                    executable.insert(task_id, inst);
+                }
             }
             StorageProvider::Dynamo => {
                 bail!("drop dynamodb keyspace is not implemented")
